🔥 chore(index.js): remove commented out code and unused h1 tag The commented out code for password reset functionality is no longer needed and has been removed. The unused h1 tag has also been removed to improve code cleanliness.
🔥 chore(resultPredict.js): remove unnecessary whitespace
The commit removes an unnecessary whitespace in the code.

🎨 style(resultPredict.js): improve UI layout and add more information
The changes in this commit improve the UI layout of the resultPredict page by adjusting the width of the table and adding a scrollbar to the table. Additionally, the commit adds more information to the page by including a section on the methodology used to predict grades and careers. This section includes information on the statistical analysis used to predict grades and careers, as well as a disclaimer that the results may not be entirely accurate due to the exclusion of certain variables. Finally, the commit also moves the "back" button to the bottom of the page for better user experience.